Restructure routes
