+++ src/auth/components/login/Login.tsx
@@ -1,14 +1,13 @@
 import React from 'react';
 
-import authenticate from '../auth/authenticate';
-import PageLayout from '../common/pageLayout/PageLayout';
+import authenticate from '../../authenticate';
+import PageLayout from '../../../common/pageLayout/PageLayout';
 
 type Props = {};
 
 function Home(props: Props) {
   return (
     <PageLayout>
-      Frontpage
       <span onClick={authenticate} role="button">
         Login using tunnistamo
       </span>
